backend/requirements.txt
fastapi>=0.103
uvicorn[standard]>=0.23
pydantic>=2.4,<3.0
python-dotenv>=1.0
openai>=1.3.0
python-multipart>=0.0.6
anthropic>=0.7.0
google-genai>=0.2.0
pydantic-settings>=2.0.0
